Update 9000-AscheAI.as
As of version 1 revision 13, Burning Ash Blast has been reworked to be weaker at close range, but more powerful at longer distances. As such, the minimum and maximum range with which the AI will use it has been increased.

The ability to use Flame of Reincarnation has also been added.